pyproject.toml
[project]
name = "wow-token-price-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=3.2.0",
    "flask-caching>=2.3.1",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.14.5",
    "schedule>=1.2.2",
]
